requirements.txt
beautifulsoup4==4.4.1
dj-database-url==0.4.1
Django==1.9.6
django-webtest==1.7.9
djangorestframework==3.3.3
gunicorn==19.6.0
mock==2.0.0
requests==2.19.1
six==1.10.0
textblob==0.15.1
waitress==0.9.0
WebOb==1.6.1
WebTest==2.0.21
